rdbms: Convert SQLite values to string
Prior to PHP 8.1, the SQLite PDO returned all values as strings or null.
Starting with PHP 8.1, the data type of data from the database is
respected. This is a breaking change and causes problems for at least
Wikibase.

To prevent these issues, we explicitly cast all non-null returned
values to strings to restore the old behavior.

Bug: T316487
Change-Id: I0098a7345835bef89be4ff86b3b91829583c571f